fix: separating interface for ProfileService
