+++ frontend/src/features/cases/services/profileService.ts
@@ -1,90 +1,7 @@
 import apiClient from '../../../shared/services/apiClient';
+import type { GenerateProfileRequest, GenerateProfileResponse } from './types/profile.types';
 
-export interface GenerateProfileFilters {
-  dateFrom?: string;
-  dateTo?: string;
-  channel?: string;
-  type?: string;
-}
 
-export interface GenerateProfileRequest {
-  tenantId: string;
-  filters?: GenerateProfileFilters;
-  notes?: string;
-}
-
-export interface DetectedAnomaly {
-  date: string;
-  type: string;
-  amount: number;
-  description: string;
-  risk: 'High' | 'Medium' | 'Low';
-}
-
-export interface ProfileMetrics {
-  totalVolume: number;
-  totalValue: number;
-  crossBorderCount: number;
-  entityId: string;
-  entityName: string;
-  entityRole: 'Creditor' | 'Debitor';
-  transactionCurrency: string;
-  transactionType: string;
-  entityType?: string;
-}
-
-export interface SummaryTable {
-  totalVolume: number;
-  totalValue: number;
-  avgTicketSize: number;
-  deviationPercent: string;
-}
-
-export interface TransactionProfile {
-  caseId: string;
-  filters?: Record<string, string | number>;
-  metrics: ProfileMetrics;
-  outliers?: Record<string, unknown>;
-  summaryTable?: SummaryTable;
-  notes?: string;
-  visualization?: string;
-  detectedAnomalies?: DetectedAnomaly[];
-}
-export interface TransactionRecordDto {
-  transaction_id: number;
-  event_date: string; // ISO date
-  tx_amount: number;
-  tx_ccy: string;
-  tx_type: string;
-  is_alerted: number;
-  is_investigated: number;
-  cum_tx_count: number;
-  cum_tx_amount: number;
-  entity_role: string;
-  entity_id: string;
-  entity_type: string;
-
-  // optional depending on query
-  debtor_name?: string;
-  creditor_name?: string;
-}
-
-export interface SqlResponseDto {
-  status: string;
-  code: number;
-  query: string;
-  row_count: number;
-  data: TransactionRecordDto[];
-}
-
-export interface GenerateProfileResponse {
-  tenantId: string;
-  transactionCreditorResp: SqlResponseDto;
-  transactionDebtorResp: SqlResponseDto;
-
-}
-
-export interface GetProfileResponse extends TransactionProfile { }
 
 export class ProfileService {
   private readonly baseUrl = '/api/v1/lakehouse';
@@ -122,17 +39,6 @@ export class ProfileService {
       throw this.handleError(error, 'generate transaction profile');
     }
   }
-
-  // async getProfile(caseId: number): Promise<GetProfileResponse> {
-  //   try {
-  //     const response = await apiClient.get<GetProfileResponse>(
-  //       `${this.baseUrl}/${caseId}`,
-  //     );
-  //     return response;
-  //   } catch (error: any) {
-  //     throw this.handleError(error, 'get transaction profile');
-  //   }
-  // }
 
   private handleError(error: any, operation: string): Error {
     if (error.response?.data) {
